pyproject.toml
[project]
name = "students-grade-management-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "cloudinary>=1.44.1",
    "django>=6.0.1",
    "django-celery-beat>=2.1.0",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "mysqlclient>=2.2.7",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
]
